requirements.txt
phonenumbers
requests
colorama
python-whois
bs4
googlesearch-python
pillow